Reinstate typedef's in the C API
The way it was, the C API would cause compiler errors, because the opaque
structs llvm_dsp and llvm_dsp_factory were defined without a typedef, but were
being referenced as "foo" instead of "struct foo". So the choice was to either
use "struct <name>" throughout, or simply switch back to using typedef.  I chose
the latter since it is easier, and because I think it is nicer to use.