Removes titling capabilities from jquery.svg.js.
SVG titles are interpreted as tooltips in FireFox. As a result, titles of
svg nodes, which are text containing functions in jquerysvg, appear
as tooltips, spewing snippets of code in yellow along with the faust-
generated tooltips.

The simple solution to this problem is to comment out the titling capabilities
from jquerysvg, as they are not used in this project.

Also ran make depend, as the hex dumps have changed folders to hexa and
this resulted in the old hexdump stuff staying around and needing manual
deletion.