Implements the possibility to create multiple UIs on a single webpage.
Instead of creating a div statically and then adding an SVG to it,
we create the div dynamically in JavaScript.  This allows for the possibility
of creating multiple SVGs in multiple divs - one for each UI description
passed into the main function. For now, because the UIs are described by
single applications, this will only be one UI per application. However,
it opens up the possibility for a protocol in which multiple applications
concurrently communicate their UI information to a single webpage that must
draw these UIs.

What would need to be implemented for the above to work is a protocol
for the layout of multiple UIs.  Currently, a naïve solution is put into
place whereby the UIs are simply stacked on top of each other.