Implements dynamic callback creation for UIs.
Allows for UI creators (servers, audio nodes, etc.) to register their callback
with the javascript listeners. This callback is then called whenever a
UI object's state changes. The callback should always be of the form:

function(dest, vale) {
  // do something
}

where dest is an address in a faust application and value is the
(assumedly new) value of the object at that address.