Enhanced -effect option.
The -effect option has been enhanced to accept an 'auto' parameter. In this case the effect used is the one corresponding to the definition of 'effect' in the main dsp file.
For example `faust2caqt -midi -effect auto foo.dsp` will use the definition of 'effect' in foo.dsp as effect.