Version 2.6.0: adds support for function level metadata.
In order to facilitate cooperation between authors of Faust libraries, in particular the fact that multiple authors can contribute to the same library, a new declare statement is introduced: `declare <function> <key> "<value>";`. It allows to declare for each function of a library its own metadata, for example its author, its copyrights, its license, etc. 

Metadata for a function can be anywhere in the file, but we recommend placing it just before the function definition.