gui: add runAllGuis() function
This function allows all GUI instances to be run, avoiding
multiple #ifdefs in the eg: jack-gtk.cpp code, and other
architecture files. Note this is currently implemented in the GUI level,
so all *GUI* subclasses #ifdefs can be removed, however the UI
subclasses must continue to use the #ifdef method.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>