jack-gtk: refactor to use GUI::runAllGuis()
This commit refactors the jack-gtk.c architecture file
to use the newly added GUI::runAllGuis() function, which
reduces the number of #ifdefs required in compiling with
specific GUIs enabled or disabled.

The OSCCTRL and MIDICTRL UIs derive from GUI, so the #idef
run() function calls are removed as they are now called
indirectly via GUI::runAllGuis().

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>