Work around a compiler bug in MSVC from Visual Studio 2019.
The bug causes the little state machine in evalLabel to "stutter" in some
cases (e.g., %k becomes %kk). Apparently that's because in Release mode the
compiler optimizes this loop so aggressively that it misses out the increment
of the src pointer when switching to state 2 after the first IdentChar is
detected.

We work around this by just dropping the local char c variable and using the
src pointer directly throughout the loop, incrementing it whenever necessary.
This solves the issue.