Only generates signed fixed types until LSB inference in loops is fixed
This is because in integer loops, LSB will be set to -24 by default for safety,
which will cause the content of the loop not to be treated like an integer and
thus not wrapping correctly (which causes problems eg in noise).
Until we have a way to infer that a loop only produces integers, we use signed
types to accomodate negative integers caused by wrapping.