create a method for making plugin_name compatible in case it is invalid (for instance if starts with a number)
Faust compiles files that may start with a number, but wwise cannot handle such project names

thesolution applied is to prefix with _Dsp keyword

otherwise we simply capitalize the plugin_name variable

note: plugin name is by default initialized from the name of the dsp file. But the user is able to name a wwise project by declaring the name as a metadata from within the dsp file. This is why setting of project_name may be overriden after being initialized.