add support for deleting installed plugins in faust2wwise test, particularly for Windows/msys platforms. Cleaning installed plugins on MacOS is yet undetermined since the installation directory cannot be reliably retrieved due to hashed keys produced in namings of the directories where the plugin files are installed.
To support deleting plugins from the installation dir, the faust2wwise configuration is now exported into a json file, making possible to access information about the installation path of the plugin from an external process. Note that faust2wwise in test mode, calls faust2wwise in operational mode for each dsp example existing within a given directory.

Note: The configuration file is exported immediately after locking, that is after step 1 (compiling the dsp file using the faust compiler). This is because the plugin configuration is finalized after having compiled the dsp file, proccessed the json file of the faust config, and retrieving information regarding the type of plugin, the number of inputs/outputs  etc.